Changing parseRequest() for checkboxes who are printed with index.
Due to the previous `getPresenceInputName()`, where _ was appended to the leaf path, the regex was wrong as the index would be printed before the _

The trailing underscore was previously used to prevent the regex from matching in the first `if`, so this was changed to ensure that it only matched from the start of the string.

We then keep track of which ones have been marked as true, to ensure we don't mark them as false afterwards.